Allows the dialog header to accept a custom ReactNode
I wanted a way to increase the title size of the header and add
additional elements, like icons. I'm suggesting an update to the dialog
component to continue rendering the string passed into the title prop as
it does today, but to also extend it so that a fully custom react node
can be passed in as well.